+++ Block/Adminhtml/GenericBackButton.php
@@ -1,0 +1,30 @@
+<?php
+declare(strict_types=1);
+
+namespace Panth\PageBuilderAi\Block\Adminhtml;
+
+use Magento\Framework\UrlInterface;
+use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
+
+class GenericBackButton implements ButtonProviderInterface
+{
+    public function __construct(
+        private readonly UrlInterface $urlBuilder
+    ) {
+    }
+
+    public function getButtonData(): array
+    {
+        return [
+            'label' => __('Back'),
+            'on_click' => sprintf("location.href = '%s';", $this->getBackUrl()),
+            'class' => 'back',
+            'sort_order' => 10,
+        ];
+    }
+
+    private function getBackUrl(): string
+    {
+        return $this->urlBuilder->getUrl('*/*/');
+    }
+}